+++ lobject.cpp
@@ -89,7 +89,7 @@ lua_Number luaO_arith (int op, lua_Number v1, lua_Number v2) {
     case LUA_OPROTR: return luai_numrotr(NULL, v1, v2);
     case LUA_OPUNM:  return luai_numunm(NULL, v1);
     case LUA_OPBNOT: return luai_numbnot(NULL, v1);
-    default: lua_assert(0); return 0;
+    default: lua_assert(0); return (int32_t)0;
   }
 }
 
@@ -165,10 +165,10 @@ static lua_Number lua_strx2number (const char *s, char **endptr) {
 #endif
 
 
-static lua_Number readany (const char **s, lua_Number r, int *count, int base, int max = INT_MAX) {
+static lua_Number readany (const char **s, lua_Number r, int *count, int32_t base, int max = INT_MAX) {
   for (; lisxdigit(cast_uchar(**s)); (*s)++, max--) {
     if (max > 0) {
-      int d = luaO_hexavalue(cast_uchar(**s));
+      int32_t d = luaO_hexavalue(cast_uchar(**s));
       if (d >= base) break;
       r = r * cast_num(base) + cast_num(d);
       (*count)++;
@@ -249,7 +249,7 @@ const char *luaO_pushvfstring (lua_State *L, const char *fmt, va_list argp) {
         break;
       }
       case 'd': {
-        setnvalue(L->top++, cast_num(va_arg(argp, int)));
+        setnvalue(L->top++, cast_num((int32_t)va_arg(argp, int)));
         break;
       }
       case 'f': {
